// logic/types/common.ts
/**
 * This file was automatically generated by joi-to-typescript
 * Do not modify this file manually
 */

export interface BaseIdInput {
  id: ObjectId;
}

export interface BaseInput {
  body?: any;
  query?: undefined;
  user?: UserSession;
}

export interface BaseOutput {
  message?: string;
  result: any;
}

export type Email = string;

export type Id = string;

export type JwtToken = string;

export type ObjectId = string;

export interface UserSession {
  email: Email;
  exp: number;
  iat: number;
  id: Id;
  name: string;
}